Extract diff printers to own files
