+++ lib/reporters/base.js
@@ -9,6 +9,8 @@ var diff = require('diff');
 var ms = require('../ms');
 var utils = require('../utils');
 var supportsColor = process.browser ? null : require('supports-color');
+var UnifiedDiff = require('./diff/unified-diff');
+var InlineDiff = require('./diff/inline-diff');
 
 /**
  * Expose `Base`.
@@ -127,6 +129,9 @@ var color = exports.color = function (type, str) {
   return open + str.replace(closeRE, open) + close;
 };
 
+var unifiedDiff = new UnifiedDiff(color, '      ');
+var inlineDiff = new InlineDiff(color, '      ');
+
 /**
  * Expose term window size, with some defaults for when stderr is not a tty.
  */
@@ -190,7 +195,7 @@ function showDiff (err) {
 exports.generateDiff = function (actual, expected) {
   var diff = doDiff(expected, actual, exports.inlineDiffs);
 
-  return exports.inlineDiffs ? inlineDiff(diff) : unifiedDiff(diff);
+  return exports.inlineDiffs ? inlineDiff.print(diff) : unifiedDiff.print(diff);
 };
 
 /**
@@ -374,111 +379,6 @@ Base.prototype.epilogue = function () {
 };
 
 /**
- * Pad the given `str` to `len`.
- *
- * @api private
- * @param {string} str
- * @param {string} len
- * @return {string}
- */
-function pad (str, len) {
-  str = String(str);
-  return Array(len - str.length + 1).join(' ') + str;
-}
-
-/**
- * Converts an inline diff between two strings to coloured text representation.
- *
- * @api private
- * @param {Object} diff structured differences actual from expected
- * @return {string} The diff for output to the console.
- */
-function inlineDiff (diff) {
-  var indent = '      ';
-  var lines = [''];
-  diff.forEach(function (change) {
-    function colorize (line) {
-      if (change.added) {
-        return color('diff added', line);
-      }
-      if (change.removed) {
-        return color('diff removed', line);
-      }
-      return line;
-    }
-
-    change.value.split('\n').forEach(function (line, i) {
-      line = colorize(line);
-      if (i === 0) {
-        line = lines.pop() + line;
-      }
-      lines.push(line);
-    });
-  });
-
-  if (lines.length > 4) {
-    var width = String(lines.length).length;
-    lines = lines.map(function (line, i) { return indent + pad(i + 1, width) + ' | ' + line; });
-  } else {
-    lines = lines.map(function (line) { return indent + line; });
-  }
-  lines.push('');
-
-  return '\n' + indent +
-    color('diff removed', 'expected') + ' ' +
-    color('diff added', 'actual') +
-    '\n\n' +
-    lines.join('\n');
-}
-
-/**
- * Converts a unified diff between two strings to coloured text representation.
- *
- * @api private
- * @param {Object} diff structured differences actual from expected
- * @return {string} The diff for output to the console.
- */
-function unifiedDiff (diff) {
-  var indent = '      ';
-
-  function inlineColorize (change) {
-    if (change.added) { return color('inline diff added', change.value); }
-    if (change.removed) { return color('inline diff removed', change.value); }
-    return change.value;
-  }
-  function colorize (line) {
-    if (line.kind === '+') {
-      let text = line.kind + line.changes.map(inlineColorize).join('');
-      return indent + color('diff added', text);
-    }
-    if (line.kind === '-') {
-      let text = line.kind + line.changes.map(inlineColorize).join('');
-      return indent + color('diff removed', text);
-    }
-    return indent + line.kind + line.value;
-  }
-
-  var lines = [];
-  diff.hunks.forEach(function (hunk, i) {
-    if (i !== 0) {
-      lines.push('--');
-    }
-    hunk.lines.forEach(function (line) {
-      if (line.kind !== '\\') {
-        lines.push(colorize(line));
-      }
-    });
-  });
-  lines.push('');
-
-  return '\n' + indent +
-    color('diff removed', '- expected') + ' ' +
-    color('diff added', '+ actual') +
-    '\n\n' +
-    lines.join('\n');
-}
-
-/**
  * Creates diffs for given error by comparing `expected` and `actual` converted to strings.
  *
  * @api private
@@ -490,7 +390,7 @@ function makeDiff (err) {
     ? err.diff(exports.inlineDiffs)
     : generateDiff(err, exports.inlineDiffs);
 
-  return exports.inlineDiffs ? inlineDiff(diff) : unifiedDiff(diff);
+  return exports.inlineDiffs ? inlineDiff.print(diff) : unifiedDiff.print(diff);
 }
 
 /**
